Refactored to use routers
